organizing service calls out of the component and into the services folder
